workspace-backend/requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
isort==7.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
pytokens==0.3.0
sqlparse==0.5.5
